Restore persistent data generation for C (#198)
Introduce --language option to enable language-specific linkage
features and generate a dictionary of globals in c4 for C programs.

- Fixed working with tvm exception

+++ tvm_linker/src/testcall.rs
@@ -382,7 +382,7 @@ pub fn call_contract_ex<F>(
             }
             _ => -1
         }
-        Ok(code) => code as i32
+        Ok(code) => code,
     };
     println!("TVM terminated with exit code {}", exit_code);
     println!("Gas used: {}", engine.get_gas().get_gas_used());
